+++ apps/ledger-live-mobile/src/newArch/features/LandingPages/screens/LargeMoverLandingPage/index.tsx
@@ -1,69 +1,113 @@
-import { Flex, InfiniteLoader, Text, Button } from "@ledgerhq/native-ui";
-import React, { useState } from "react";
+import { Flex } from "@ledgerhq/native-ui";
+import React, { useState, useMemo } from "react";
 import { useLargeMover } from "./hooks/useLargeMover";
 import getWindowDimensions from "~/logic/getWindowDimensions";
 import { Card } from "./components/Card";
 import { LandingPagesNavigatorParamList } from "~/components/RootNavigator/types/LandingPagesNavigator";
-import { ScreenName } from "~/const";
+import { NavigatorName, ScreenName } from "~/const";
 import { StackNavigatorProps } from "~/components/RootNavigator/types/helpers";
 import { StickyHeader } from "./components/StickyHeader";
 import { SafeAreaView } from "react-native";
 import { useTheme } from "styled-components/native";
+import { getCurrencyIdsFromTickers, rangeMap } from "./utils";
+import { SwiperComponent } from "~/newArch/components/Swiper/components/Swiper";
+import { NavigationProp, useNavigation } from "@react-navigation/native";
+import { WalletTabNavigatorStackParamList } from "~/components/RootNavigator/types/WalletTabNavigator";
+import { LoadingIndicator } from "./components/Loading";
+import { CardType } from "./types";
+import { TrackScreen } from "~/analytics";
+import { PAGE_NAME } from "./const";
 
+import { useSelector } from "react-redux";
+import { useLargeMoverChartData } from "@ledgerhq/live-common/market/hooks/useLargeMoverChartData";
+import { counterValueCurrencySelector } from "~/reducers/settings";
+import { KeysPriceChange } from "@ledgerhq/live-common/market/utils/types";
 type LargeMoverLandingPageProps = StackNavigatorProps<
   LandingPagesNavigatorParamList,
   ScreenName.LargeMoverLandingPage
 >;
 
 export const LargeMoverLandingPage = ({ route }: LargeMoverLandingPageProps) => {
-  const { currencyIds, initialRange } = route.params;
-  const { range, setRange, currencies, loading } = useLargeMover({
-    currencyIds,
-    initialRange,
+  const { currencyIds, initialRange = "day" } = route.params;
+  const counterValueCurrency = useSelector(counterValueCurrencySelector);
+  const currencyIdsArray = currencyIds.toUpperCase().split(",");
+  const currenciesIds = getCurrencyIdsFromTickers(currencyIdsArray);
+  const [range, setRange] = useState<KeysPriceChange>(rangeMap[initialRange]);
+
+  const { currencies, loading, isError } = useLargeMover({
+    currenciesIds,
   });
+
+  const { chartDataArray, loadingChart } = useLargeMoverChartData({
+    ids: currenciesIds,
+    counterCurrency: counterValueCurrency.ticker,
+    range,
+  });
+
+  const navigation = useNavigation<NavigationProp<WalletTabNavigatorStackParamList>>();
 
   const { colors } = useTheme();
   const height = getWindowDimensions().height * 0.75;
-
   const [currentIndex, setCurrentIndex] = useState(0);
 
-  const handleNext = () => {
-    setCurrentIndex(prevIndex => (prevIndex + 1) % currencies.length);
+  const currenciesWithId = useMemo(
+    () =>
+      currencies.map((currency, index) => {
+        const chartDataEntry = chartDataArray.find(
+          chartData => chartData.idChartData === currency.id,
+        );
+        return {
+          id: currency.id,
+          data: currency.data,
+          chartData: chartDataEntry?.chartData,
+          loading: loadingChart,
+          idCard: index,
+        };
+      }),
+    [chartDataArray, currencies, loadingChart],
+  );
+  const renderCard = (card: CardType) => {
+    if (!card.data) return null;
+
+    const chartDataEntry = chartDataArray.find(c => c.idChartData === card.id);
+
+    return (
+      <Card
+        data={card.data}
+        chartData={chartDataEntry?.chartData}
+        loading={loadingChart}
+        height={height}
+        range={range}
+        setRange={setRange}
+      />
+    );
   };
 
-  if (loading) {
-    return (
-      <Flex height={height} width="100%" justifyContent="center" alignItems="center">
-        <InfiniteLoader color="primary.c50" size={38} />
-      </Flex>
-    );
-  }
-
-  if (currencies.length === 0) {
-    return (
-      <Flex>
-        <Text>{"No data available"}</Text>
-      </Flex>
-    );
+  if (isError) {
+    navigation.navigate(NavigatorName.Market, {
+      screen: ScreenName.MarketList,
+      params: { top100: true },
+    });
   }
 
   return (
-    <SafeAreaView style={{ flex: 1, backgroundColor: colors.neutral.c00 }}>
-      <Flex width="100%" flex={1}>
-        <StickyHeader />
-        <Button onPress={handleNext} size="large">
-          {"Next"}
-        </Button>
-        <Flex padding={1}>
-          {currencies[currentIndex].data && currencies[currentIndex].chartData ? (
-            <Card
-              {...currencies[currentIndex].data!}
-              chartData={currencies[currentIndex].chartData!}
-              range={range}
-              setRange={setRange}
+    <SafeAreaView style={{ flex: 1, backgroundColor: colors.neutral.c00, paddingTop: 10 }}>
+      <TrackScreen name={PAGE_NAME} initialRange={initialRange} currencyIds={currencyIds} />
+      <StickyHeader />
+      <Flex paddingTop={40}>
+        {loading && loadingChart ? (
+          <LoadingIndicator height={height} />
+        ) : (
+          <Flex>
+            <SwiperComponent
+              cardContainerStyle={{ justifyContent: "flex-start" }}
+              currentIndex={currentIndex}
+              onIndexChange={setCurrentIndex}
+              initialCards={currenciesWithId}
+              renderCard={renderCard}
             />
-          ) : null}
-        </Flex>
+          </Flex>
+        )}
       </Flex>
     </SafeAreaView>
   );