+++ apps/ledger-live-mobile/src/newArch/features/LandingPages/screens/LargeMoverLandingPage/types.d.ts
@@ -13,4 +13,10 @@ type InformationsProps = {
   allTimeLowDate: Date;
 };
 
-export type { InformationsProps };
+type CardType = {
+  id: string;
+  data?: CurrencyData | undefined;
+  chartData?: MarketCoinDataChart | undefined;
+};
+
+export type { InformationsProps, CardType };