+++ libs/coin-modules/coin-aptos/src/types/index.ts
@@ -69,6 +69,25 @@ export type TransactionRaw = TransactionCommonRaw & {
   errors?: string;
 };
 
+export type AptosFungibleStoreResourceData = {
+  balance: BigNumber;
+  frozen: boolean;
+  metadata: { inner: string };
+};
+
+export type AptosFungibleoObjectCoreResourceData = {
+  owner: string;
+};
+
+export type AptosMoveResourceData = {
+  guid: { id: { addr: string; creation_num: string } };
+};
+
 export type AptosMoveResource = {
-  [key: string]: { guid: { id: { addr: string; creation_num: string } } };
+  [key: string]: AptosMoveResourceData;
+};
+
+export type AptosBalance = {
+  asset_type: string;
+  amount: BigNumber;
 };